remarks fix
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@3142 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ top/modules/TOPReconstruction/src/TOPreco.cc
@@ -188,7 +188,7 @@ namespace Belle2 {
 
     double TOPreco::PDF(int chID, double T, double Mass)
     {
-      float t = T; float mass = Mass;
+      float t = (float) T; float mass = (float) Mass;
       return get_pdf_(&chID, &t, &mass);
     }
